Move tests to acceptance
