+++ web/tests/acceptance/authenticated/document-test.ts
@@ -1,4 +1,4 @@
-import { visit } from "@ember/test-helpers";
+import { click, findAll, visit } from "@ember/test-helpers";
 import { setupApplicationTest } from "ember-qunit";
 import { module, test } from "qunit";
 import { authenticateSession } from "ember-simple-auth/test-support";
@@ -29,5 +29,87 @@ module("Acceptance | authenticated/document", function (hooks) {
     });
     await visit("/document/1?draft=true");
     assert.equal(getPageTitle(), "Test Document | Hermes");
+  });
+
+  test("you can change a draft's product area", async function (this: AuthenticatedDocumentRouteTestContext, assert) {
+    const docID = "test-doc-0";
+
+    this.server.createList("product", 3);
+
+    const initialProduct = this.server.schema.products.find(2).attrs;
+
+    const initialProductName = initialProduct.name;
+    const initialProductAbbreviation = initialProduct.abbreviation;
+
+    const targetProductAbbreviation =
+      this.server.schema.products.find(1).attrs.abbreviation;
+
+    this.server.create("document", {
+      objectID: docID,
+      isDraft: true,
+      product: initialProductName,
+    });
+
+    await visit(`/document/${docID}?draft=true`);
+
+    const docNumberSelector = "[data-test-sidebar-doc-number]";
+    const productSelectSelector = "[data-test-sidebar-product-select]";
+    const productSelectTriggerSelector = "[data-test-badge-dropdown-trigger]";
+    const productSelectDropdownItemSelector =
+      "[data-test-product-select-badge-dropdown-item]";
+
+    assert
+      .dom(docNumberSelector)
+      .hasText(initialProductAbbreviation, "The document number is correct");
+
+    assert
+      .dom(productSelectSelector)
+      .exists("drafts show a product select element")
+      .hasText(initialProductName, "The document product is selected");
+
+    await click(productSelectTriggerSelector);
+    const options = findAll(productSelectDropdownItemSelector);
+
+    const expectedProducts = [
+      "Test Product 0",
+      "Test Product 1",
+      "Test Product 2",
+    ];
+
+    options.forEach((option: Element, index: number) => {
+      assert.equal(
+        option.textContent?.trim(),
+        expectedProducts[index],
+        "the product is correct"
+      );
+    });
+
+    await click(productSelectDropdownItemSelector);
+
+    assert
+      .dom(docNumberSelector)
+      .hasText(
+        targetProductAbbreviation,
+        "The document is patched with the correct docNumber"
+      );
+
+    this.server.schema.document
+      .findBy({ objectID: docID })
+      .update("isDraft", false);
+  });
+
+  test("a published doc's productArea can't be changed ", async function (this: AuthenticatedDocumentRouteTestContext, assert) {
+    this.server.create("document", {
+      objectID: 1,
+      title: "Test Document",
+      isDraft: false,
+      product: "Test Product 0",
+    });
+
+    await visit("/document/1");
+
+    assert
+      .dom("[data-test-sidebar-product-select]")
+      .doesNotExist("published docs don't show a product select element");
   });
 });
